OeX_ES-33: Refactoring for es documents. Moves all common fields to Base document.
